refactor(WebAPI): remove mediatr code and registered replacement pipelines
