+++ src/Application/Config/Autofac/FastEndpointsModule.cs
@@ -1,4 +1,7 @@
+using System.Reflection;
 using Autofac;
+using FluentValidation;
+using Module = Autofac.Module;
 
 namespace PlexRipper.Application;
 
@@ -8,6 +11,21 @@ public class FastEndpointsModule : Module
     {
         builder.RegisterType<CommandExecutor>().As<ICommandExecutor>().InstancePerLifetimeScope();
 
-        builder.RegisterType<EventPublisher>().As<IEventPublisher>().InstancePerLifetimeScope();
+        builder.RegisterType<EventPublisher>().As<IEventPublisher>().SingleInstance();
+
+        var assembly = Assembly.GetExecutingAssembly();
+
+        // register all I*Commands
+        builder
+            .RegisterAssemblyTypes(assembly)
+            .Where(t => t.Name.EndsWith("Command"))
+            .AsImplementedInterfaces()
+            .SingleInstance();
+
+        // register all FluentValidators
+        builder
+            .RegisterAssemblyTypes(assembly)
+            .Where(t => t.IsClosedTypeOf(typeof(IValidator<>)))
+            .AsImplementedInterfaces();
     }
 }
