+++ src/Domain/GlobalUsings.cs
@@ -4,5 +4,4 @@ global using System.ComponentModel.DataAnnotations;
 global using System.ComponentModel.DataAnnotations.Schema;
 global using FluentResults;
 global using Logging;
-global using MediatR;
 global using PlexRipper.Domain;
